Rename octave-head to octave-default and don't use version "HEAD"
Using a version named "HEAD" has special meaning to brew, and will
break the installation with an "empty installation" error if you
try to use it in a normal formula.

Renamed "octave-head" to "octave-default" to reflect the Octave
maintainers' convention of using "default" as the main development
branch. Corresponds better with "octave-stable".